cleaning

+++ jump_table.h
@@ -22,7 +22,7 @@
     }
 
 #define JUMP_MENU_STRINGIFY(menu, dest) \
-    int dest##_remaining = sizeof(dest);\
+    size_t dest##_remaining = sizeof(dest);\
     for (int index = 0; index < JUMP_TABLE_COUNT(menu) && dest##_remaining > 0; index++) { \
         char* start = (dest + sizeof(dest)) - dest##_remaining; \
         dest##_remaining -= snprintf(start, dest##_remaining, "%d: %s\n", index, menu[index].desc ); \
